# Documentation/Open API документация/OpenApi.yaml
openapi: 3.0.3
info:
  title: Сервис для создания и управления портфолио Artfolio
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  version: 0.0.1
servers:
  - url: http://79.174.93.30:8080
  - url: http://79.174.93.30
  - url: http://localhost:8080
...
paths:
...
  /api/v1/feed:
    get:
      tags:
        - feed
      summary: Получить страницу постов
      description: Позволяет пользователю получить страницу постов определенной секции (Новые, Популярные, Для вас). Популярные посты определяются по количеству лайков. Секция "Для вас" содержит посты в первую очередь тех, на кого пользователь подписан.
      parameters:
        - name: section
          description: Указывается конкретная секция NEW (по-умолчанию), POPULAR, SUBSCRIBE
          in: query
          required: false
          schema:
            type: string
            enum:
              - NEW
              - POPULAR
              - SUBSCRIBE
        - name: _page
          description: Номер страницы (отчет от 0)
          in: query
          required: true
          schema:
            type: integer
        - name: _limit
          description: Максимальное количество элементов на одной странице (по-умолчанию 20)
          in: query
          required: false
          schema:
            type: integer
      operationId: getSectionPosts
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostsPage'
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/feed/search:
    get:
      tags:
        - feed
      summary: Поиск постов по параметрам
      description: Позволяет пользователю получить страницу постов по названию
      parameters:
        - name: name
          in: query
          required: true
          schema:
            type: string
        - name: _page
          description: Номер страницы (отчет от 0)
          in: query
          required: true
          schema:
            type: integer
        - name: _limit
          description: Максимальное количество элементов на одной странице (по-умолчанию 20)
          in: query
          required: false
          schema:
            type: integer
      operationId: getPostsByName
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostsPage'
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  # user
...
  /api/v1/user/{username}/subscribes:
    get:
      tags:
        - subscribes
      summary: Получить тех пользователей, на кого пользователь c username подписан
      description: Позволяет получить страницу пользователей, на кого пользователь с username подписан
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
      operationId: getAllUserSubscribes
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsersPage'
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      tags:
        - subscribes
      summary: Подписаться на пользователя c username
      description: Позволяет подписаться на пользователя (username)
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
      operationId: subscribe
      responses:
        '200':
          description: Successful operation
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
        - bearerAuth: [ ]
    delete:
      tags:
        - subscribes
      summary: Отписаться от пользователя с username
      description: Позволяет отписаться от пользователя с username
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
      operationId: deleteSubscribe
      responses:
        '200':
          description: Successful operation
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
        - bearerAuth: [ ]
  # user followers
...
  /api/v1/auth:
    get:
      tags:
        - auth
      summary: Получить данные пользователя по JWT токену
      description: Позволяет по JWT токену в Header получить данные пользователя
      operationId: getUserDataByJwt
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
        - bearerAuth: [ ]
  /api/v1/auth/login:
    post:
      tags:
        - auth
      summary: Войти в аккаунт (получить JWT токен)
      description: Позволяет зарегистрированному пользователю получить JWT bearer token
      operationId: login
      requestBody:
        description: Данные пользователя для входа
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthRequest'
        required: true

      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/auth/register:
    post:
      tags:
        - auth
      summary: Зарегистрировать аккаунт (получить JWT токен)
      description: Позволяет анонимному пользователю получить JWT bearer token, а также зарегистрироваться
      operationId: register
      requestBody:
        description: Данные пользователя для входа
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
            encoding:
              userInfo:
                contentType: application/json
        required: true
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/auth/change-password:
    patch:
      tags:
        - auth
      summary: Сбросить и изменить пароль по Секретному слову
      description: Позволяет зарегистрированному пользователю изменить пароль по секретному слову
      requestBody:
        description: Данные пользователя для сброса пароля
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordRequest'
        required: true
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
components:
  schemas:
...
    RegisterRequest:
      type: object
      properties:
        userInfo:
          type: object
          properties:
            fullName:
              type: string
              example: Рамси Болтон
            description:
              type: string
              example: Родился я бастардом на Севере Вестероса. С раннего детства отец заметил мой талант к Искусству. Пока я новичок в этом деле. Надеюсь вам понравятся мои работы.
            country:
              type: string
              example: Россия
            city:
              type: string
              example: Уфа
            username:
              type: string
              example: boltonArts
            email:
              type: string
              example: bolton@vesteros.com
            password:
              type: string
              example: somePassword19
            secretWord:
              type: string
              example: winterIsComing
          description: Данные пользователя в формате JSON
        avatarFile:
          type: string
          format: binary
          description: Изображение аватара в формате png/jpg
    AuthRequest:
      type: object
      properties:
        email:
          type: string
          example: bolton@vesteros.com
        password:
          type: string
          example: somePassword19
          format: password
    AuthResponse:
      type: object
      properties:
        username:
          type: string
          example: boltonArts
        fullName:
          type: string
          example: Рамси Болтон
        email:
          type: string
          example: bolton@vesteros.com
        role:
          type: string
          enum:
            - USER
            - ADMIN
        token:
          type: string
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c
    ChangePasswordRequest:
      type: object
      properties:
        email:
          type: string
        secretWord:
          type: string
        newPassword:
          type: string
    # post
...
    PostResponse:
      type: object
      properties:
        id:
          type: integer
          example: 33
        name:
          type: string
          example: Герб Болтона
        likeCount:
          type: integer
          example: 39
        previewMedia:
          deprecated: true
          type: integer
          example: 949
          description: id на превьюшку изображения (DEPRECATED) использовать отдельный endpoint {postId}/preview
        owner:
          $ref: '#/components/schemas/UserResponse'
    PostsPage:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/PostResponse'
        totalElements:
          type: integer
          example: 200
        totalPages:
          type: integer
          example: 10
    # comment
...
    UserResponse:
      type: object
      properties:
        uuid:
          type: string
          example: 550e8400-e29b-41d4-a716-446655440000
        fullName:
          type: string
          example: Рамси Болтон
        username:
          type: string
          example: boltonArts
        email:
          type: string
          example: bolton@vesteros.com
    UsersPage:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/UserResponse'
        totalElements:
          type: integer
          example: 100
        totalPages:
          type: integer
          example: 10
    # post report
...
    Error:
      type: object
      properties:
        message:
          type: string
          example: Something went wrong.
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT